+++ replacer-frontend/src/app/api/models/public-ip.ts
@@ -1,9 +1,0 @@
-/* tslint:disable */
-/* eslint-disable */
-
-/**
- * Public IP of the application used to perform the editions in Wikipedia
- */
-export interface PublicIp {
-  ip: string;
-}