+++ replacer-frontend/src/app/api/services.ts
@@ -3,4 +3,3 @@ export { ReviewApiService } from './services/review-api.service';
 export { PagesApiService } from './services/pages-api.service';
 export { DumpIndexingApiService } from './services/dump-indexing-api.service';
 export { ReplacementApiService } from './services/replacement-api.service';
-export { AdministrationApiService } from './services/administration-api.service';