+++ replacer-backend/replacer-web/src/main/java/es/bvalero/replacer/page/save/ReviewedReplacementDto.java
@@ -3,12 +3,14 @@ package es.bvalero.replacer.page.save;
 import static io.swagger.v3.oas.annotations.media.Schema.RequiredMode.NOT_REQUIRED;
 import static io.swagger.v3.oas.annotations.media.Schema.RequiredMode.REQUIRED;
 
+import com.fasterxml.jackson.annotation.JsonInclude;
 import io.swagger.v3.oas.annotations.media.Schema;
 import javax.validation.constraints.NotNull;
 import lombok.Data;
 import org.springframework.lang.Nullable;
 
 @Schema(description = "Reviewed replacement", name = "ReviewedReplacement")
+@JsonInclude(JsonInclude.Include.NON_NULL)
 @Data
 class ReviewedReplacementDto {